Added check to make sure post type is shop_coupon.
Code was running anytime save_post was being called. Added check for post type being shop_coupon so that it wouldn't run every time that save_post was being called.

+++ misc/affiliatewp-store-affiliate-id-on-save-post.php
@@ -10,30 +10,35 @@
  
 // This function is a copy of the store_discount_affiliate function from the class-woocommerce.php file in Affiliate WP.
 function jt_affwp_store_discount_affiliate( $coupon_id = 0 ) {
-
-		if( empty( $_POST['user_name'] ) ) {
-
-			delete_post_meta( $coupon_id, 'affwp_discount_affiliate' );
-			return;
-
-		}
-
-		if( empty( $_POST['user_id'] ) && empty( $_POST['user_name'] ) ) {
-			return;
-		}
-
-		if( empty( $_POST['user_id'] ) ) {
-			$user = get_user_by( 'login', $_POST['user_name'] );
-			if( $user ) {
-				$user_id = $user->ID;
+		
+		global $post;
+		
+		// Check post type and only run if type is 'shop_coupon'.
+		if( $post->post_type != 'shop_coupon' ) {
+			if( empty( $_POST['user_name'] ) ) {
+	
+				delete_post_meta( $coupon_id, 'affwp_discount_affiliate' );
+				return;
+	
 			}
-		} else {
-			$user_id = absint( $_POST['user_id'] );
+	
+			if( empty( $_POST['user_id'] ) && empty( $_POST['user_name'] ) ) {
+				return;
+			}
+	
+			if( empty( $_POST['user_id'] ) ) {
+				$user = get_user_by( 'login', $_POST['user_name'] );
+				if( $user ) {
+					$user_id = $user->ID;
+				}
+			} else {
+				$user_id = absint( $_POST['user_id'] );
+			}
+	
+			$affiliate_id = affwp_get_affiliate_id( $user_id );
+	
+			update_post_meta( $coupon_id, 'affwp_discount_affiliate', $affiliate_id );
 		}
-
-		$affiliate_id = affwp_get_affiliate_id( $user_id );
-
-		update_post_meta( $coupon_id, 'affwp_discount_affiliate', $affiliate_id );
 	}
 
 // Store affiliate ID on save_post -> function in class-woocommerce.php is storing id on woocommerce_coupon_options_save.
